CHeck if post install script returns report object.
